Try creating separate logic for breakdowns
